# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Account Service API
  version: 0.1.0
paths:
  /api/accounts/providers:
    post:
      summary: Register provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceProvider'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProvider'
  /api/accounts/providers/{id}:
    get:
      summary: Get provider by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProvider'
    put:
      summary: Update provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceProvider'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProvider'
components:
  schemas:
    ServiceProvider:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        businessName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: USA
        profilePictureUrl:
          type: string
          format: uri
        bio:
          type: string
        locationType:
          type: string
          enum: [WORKSHOP, MOBILE, BOTH]
        operatingHours:
          type: object
          additionalProperties:
            type: string
        coverageAreaZipcodes:
          type: array
          items:
            type: string
        isIdentityVerified:
          type: boolean
        isAgreedTerms:
          type: boolean
        isActive:
          type: boolean
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - fullName
        - locationType